requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
mercantile==1.2.1
multidict==6.0.5
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycairo==1.26.0
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.1.1
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
